requirements.txt
rembg[cpu]
pillow
ttkbootstrap
pyinstaller